Release v2.8.0

+++ board/zhihe/common/script/fastboot_images.bat
@@ -1,40 +1,48 @@
 :: Script for flashing images using fastboot
 
-::@echo off
+@echo off
 
-echo ###### Start the flashing tool
+fastboot getvar product 2> getvar.tmp
+find "product: a2" "getvar.tmp" >nul
 
-IF EXIST "bootzero-rvbl.bin" (
-    fastboot flash ram bootzero-rvbl.bin || goto :error
-    fastboot reboot
-    fastboot flash ram spl-with-fit-rvbl.bin || goto :error
-    fastboot reboot 
+IF %errorlevel% equ 0 (
+    echo ###### Start flash images ######
 ) ELSE (
-    fastboot flash ram emmc_boot-loader.img || goto :error
-    fastboot reboot 
+	IF EXIST "bootzero-rvbl.bin" (
+		echo ###### Load flashing tool bootzero and uboot
+		fastboot flash ram bootzero-rvbl.bin || goto :error
+		fastboot reboot
+		fastboot flash ram spl-with-fit-rvbl.bin || goto :error
+		fastboot reboot 
+	) ELSE (
+		echo ###### Load flashing tool uboot
+		fastboot flash ram spl-with-fit-rvbl.bin || goto :error
+		fastboot reboot 
+	)
+	echo ###### Wait for the flashing tool to be ready
+	ping 127.0.0.1 -n 5 >nul
 )
-
-echo ###### Wait for the flashing tool to be ready
-ping 127.0.0.1 -n 5 >nul
 
 echo ###### Flash gpt
 fastboot flash gpt emmc-gpt_primary.img || goto :error
 echo ###### Flash loader
 fastboot flash mmc0boot0 emmc_boot-loader.img || goto :error
+echo ###### Flash uboot_env
+fastboot flash uboot_env emmc-uboot_env.img || goto :error
 echo ###### Flash partition boot
 fastboot flash boot emmc-boot_a.img || goto :error
 echo ###### Flash partition system
 fastboot flash system emmc-system_a.img || goto :error
 echo ###### Flash partition app
 fastboot flash app emmc-app_a.img || goto :error
-echo ###### Flash partition home
-fastboot flash home emmc-home.img || goto :error
+echo ###### Flash partition data
+fastboot flash data emmc-data.img || goto :error
 
-echo ###### Images flashed success
+echo ###### Images flashed success ######
 pause
 exit /b 0
 
 :error
-echo ###### Images flashing failed
+echo ###### Images flashing failed ######
 pause
 exit /b 1
